+++ lib/plugins/tag/post_path.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const { encodeURL } = require('hexo-util');
-const url_for = require('../../plugins/helper/url_for');
+const url_for = require('../helper/url_for');
 
 /**
  * Post path tag
